Added shutdown request to monitor
The host can now request a shutdown by writing the appropriate
data to the MONITOR_CMD register. After REG_MONITOR_HALT_MINUTES
have past, the monitor will remove power from the host (unless the
host shuts off its own power first). The power-cape will return
a status code in MONITOR_CMD to show if the command worked or has
failed.

The purpose of this feature is to be able to ensure that a
planned shutdown actually works as expected.