disable twi slave during reboot
I've observed an issue where the i2c interface stops working.

Although I haven't found the root of the problem, it may be that
toggling the twi enable will cause the i2c to recover. By adding
twi_slave_stop() to the reboot we will have a better chance for this
to occur.