Add TIME_SYNCD bit to status register
The TIME_SYNCD bit is set true whenever REG_SECONDS_3 is written
from the twi interface. It is false on boot and can also be cleared
by a write to REG_STATUS.

The purpose of this bit is to indicate that the cape's rtc has been
syncronized to the correct date/time.